requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.128.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
humanfriendly==10.0
idna==3.11
imageio-ffmpeg==0.6.0
mpmath==1.3.0
numpy==2.4.1
onnxruntime==1.23.2
packaging==25.0
piper-tts==1.3.0
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyreadline3==3.5.4
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
soundfile==0.13.1
starlette==0.50.0
sympy==1.14.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
